fix: CRÍTICO - Resolver validation error delay float em AGnO tools
- Corrigir send_text_message delay: Optional[int] → Optional[float]
- AGnO Agent passa valores float (1.5, 0.8) mas tools esperavam int
- Resolver erro: "Input should be a valid integer, got a number with a fractional part"
- Validação: delay aceita 1.5s, 0.8s, 2.3s corretamente
- Tools WhatsApp agora compatíveis com delays naturais AGnO

+++ agente/tools/whatsapp/send_text_message.py
@@ -16,7 +16,7 @@ from ...core.tool_monitoring import monitor_tool
 async def send_text_message(
     text: str,
     phone: Optional[str] = None,
-    delay: Optional[int] = None
+    delay: Optional[float] = None
 ) -> Dict[str, Any]:
     """
     Envia mensagem de texto via WhatsApp com simulação de digitação natural.
